new version

+++ framework/v0.1.0/lib/cmd/framework/start.js
@@ -1,8 +1,9 @@
 'use strict';
 var fs      = require('fs');
-var help    = require( getPath('gina') + '/utils/helper');
+var help    = require( getPath('gina').root + '/utils/helper');
 var child   = require('child_process');
-var lib     = require( getPath('gina.lib') );
+var lib     = require( getPath('gina').lib );
+// var helpers     = require( getPath('gina').helpers );
 var console = lib.logger;
 /**
  * Framework start command - Needs to be launched as [sudo], [admin] or [root]
@@ -29,19 +30,14 @@ function Start(opt){
 
     var init = function(opt){
 
-        self.pid = opt.pid;
-        self.projects = require(GINA_HOMEDIR + '/projects.json');
-        self.services = [];
-        self.bundles = [];
+        self.pid        = opt.pid;
+        self.projects   = require(GINA_HOMEDIR + '/projects.json');
+        self.services   = [];
+        self.bundles    = [];
 
         cleanPIDs();
         console.info('Framework ready for connections');
 
-        //var project
-
-        //for (var p in self.projects) {
-        //    self.bundles.push()
-        //}
         var frameworkPath = getPath('framework');
 
 //        setUpLogs( function(err){
@@ -72,18 +68,38 @@ function Start(opt){
     var cleanPIDs = function() {
         var f = 0
             , path = _(GINA_RUNDIR + '/gina')
-            , files = fs.readdirSync( path );
+            , files = null
+        ;
 
-        for (;f < files.length; ++f) {
-            if (files[f] != self.pid) {
-                try {
-                    new _(path +'/'+ files[f]).rmSync()
-                } catch(err) {
-                    console.error(err.stack||err.message);
-                    process.exit(1);
+        if ( fs.existsSync( path ) ) {
+            try {
+
+                files = fs.readdirSync( path );
+
+            } catch (err) {
+
+                //if (err.stack)
+                //    err.stack = 'Possible write privilege exception: grant access to your user or run the command as sudo/admin\n'+ err.stack;
+
+                console.crit(err.stack||err.message);
+                process.exit(1);
+            }
+
+
+            for (;f < files.length; ++f) {
+                if (files[f] != self.pid) {
+                    try {
+                        new _(path +'/'+ files[f]).rmSync()
+                    } catch(err) {
+                        console.error(err.stack||err.message);
+                        process.exit(1);
+                    }
                 }
             }
         }
+
+
+
     }
 
     /**
@@ -94,9 +110,6 @@ function Start(opt){
         return ( typeof(self.bundles[bundle]) != 'undefined') ? true : false
     };
 
-//    var isRunning = fucntion(service) {
-//
-//    };
 
     var startAllServices = function() {
         for (var service in self.servicesList) {
